Changes to wording/fixes in grammar and spelling (#147)
* Update index.md

Basic grammatical and spelling corrections and some added information

Signed-off-by: Jaxx7594 <101913901+Jaxx7594@users.noreply.github.com>

* Update index.md

Slight change. Just made a sentence read a bit better.

Signed-off-by: Jaxx7594 <101913901+Jaxx7594@users.noreply.github.com>

* Update index.md

Slight touchup to a sentence. Specified that it is the Android release, to avoid misinterpetation by users coming from Windows etc in which it is ALSO called the Roblox Player.

Signed-off-by: Jaxx7594 <101913901+Jaxx7594@users.noreply.github.com>

* Update index.md

Corrected a spelling error I didn't catch

Signed-off-by: Jaxx7594 <101913901+Jaxx7594@users.noreply.github.com>

---------

Signed-off-by: Jaxx7594 <101913901+Jaxx7594@users.noreply.github.com>